refactor(demo): split sections into per-feature files; clean up style for v6.3.0
